22/05/2023

+++ app/src/main/java/br/senai/jandira/sp/zerowastetest/models/modelretrofit/modelAPI/NewCatadorFisico.kt
@@ -7,7 +7,7 @@ data class NewCatadorFisico(
     var telefone: String = "",
     var email: String = "",
     var senha: String = "",
-    var materiais: List<Long?>? = null,
+    var materiais: List<Int?>? = null,
 
     var cpf: String? = null,
